+++ Sources/ArknightsClient/Models/LaunchOptions.swift
@@ -53,6 +53,7 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 	var usesHighResolutionMode: Bool = true
 	var usesMetalPerformanceHUD: Bool = false
 	var usesGameMode: Bool = false
+	var usesPreciseScrolling: Bool = true
 
 	static let `default` = GameLaunchOptions(
 		displayMode: .windowed,
@@ -60,7 +61,8 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		usesGameSettings: true,
 		usesHighResolutionMode: true,
 		usesMetalPerformanceHUD: false,
-		usesGameMode: false
+		usesGameMode: false,
+		usesPreciseScrolling: true
 	)
 
 	private enum CodingKeys: String, CodingKey {
@@ -70,6 +72,7 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		case usesHighResolutionMode
 		case usesMetalPerformanceHUD
 		case usesGameMode
+		case usesPreciseScrolling
 	}
 
 	init(
@@ -78,7 +81,8 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		usesGameSettings: Bool = true,
 		usesHighResolutionMode: Bool = true,
 		usesMetalPerformanceHUD: Bool = false,
-		usesGameMode: Bool = false
+		usesGameMode: Bool = false,
+		usesPreciseScrolling: Bool = true
 	) {
 		self.displayMode = displayMode
 		self.resolution = resolution
@@ -86,6 +90,7 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 		self.usesHighResolutionMode = usesHighResolutionMode
 		self.usesMetalPerformanceHUD = usesMetalPerformanceHUD
 		self.usesGameMode = usesGameMode
+		self.usesPreciseScrolling = usesPreciseScrolling
 	}
 
 	init(from decoder: any Decoder) throws {
@@ -112,6 +117,11 @@ struct GameLaunchOptions: Codable, Sendable, Equatable {
 				Bool.self,
 				forKey: .usesGameMode
 			) ?? false
+		usesPreciseScrolling =
+			try container.decodeIfPresent(
+				Bool.self,
+				forKey: .usesPreciseScrolling
+			) ?? true
 	}
 
 	/// Unity standalone-player arguments supported by the Windows client.